@@ -7,17 +7,17 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
-import javax.annotation.Resource;
-import javax.ejb.ActivationConfigProperty;
-import javax.ejb.MessageDriven;
-import javax.inject.Inject;
-import javax.jms.Destination;
-import javax.jms.JMSContext;
-import javax.jms.Message;
-import javax.jms.MessageListener;
-import javax.jms.Queue;
-import javax.jms.TextMessage;
-import javax.transaction.Transactional;
+import jakarta.annotation.Resource;
+import jakarta.ejb.ActivationConfigProperty;
+import jakarta.ejb.MessageDriven;
+import jakarta.inject.Inject;
+import jakarta.jms.Destination;
+import jakarta.jms.JMSContext;
+import jakarta.jms.Message;
+import jakarta.jms.MessageListener;
+import jakarta.jms.Queue;
+import jakarta.jms.TextMessage;
+import jakarta.transaction.Transactional;
 
 import org.gmagnotta.jaxb.TopOrdersResponse;
 import org.gmagnotta.jaxb.TopValue;
@@ -32,7 +32,7 @@ import org.gmagnotta.jaxb.TopItemsResponse;
 
 @MessageDriven(name = "TopItemProcessor", activationConfig = {
 		@ActivationConfigProperty(propertyName = "destinationLookup", propertyValue = "java:global/remoteContext/getTopItemsCommand"),
-		@ActivationConfigProperty(propertyName = "destinationType", propertyValue = "javax.jms.Queue")
+		@ActivationConfigProperty(propertyName = "destinationType", propertyValue = "jakarta.jms.Queue")
 })
 public class TopItemProcessor implements MessageListener {
 
