@@ -2,10 +2,10 @@ package com.mycompany.app.servlet;
 import java.io.IOException;
 import java.math.BigDecimal;
 
-import javax.ejb.EJB;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import jakarta.ejb.EJB;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 
 import com.mycompany.app.service.ItemService;
 import com.mycompany.model.Item;
